+++ cbat/__init__.py
@@ -1,2 +1,2 @@
-from cbat.__main__ import *
 import cbat.models
+from cbat.main import *
